unify uid/gid

+++ app-core/util-linux/util-linux-2.41.20260401.ebuild
@@ -195,5 +195,11 @@ src_configure() {
 
 src_install() {
 	meson_src_install
+	if use uuidd ; then
+		cat > "${ED}"/usr/lib/sysusers.d/uuidd.conf <<- EOF || die
+			g uuidd 109 - -
+			u uuidd 109:109 "UUID generator helper daemon" /var/lib/libuuid /usr/bin/nologin
+		EOF
+	fi
 	rm "${ED}"/usr/share/man/man8/vigr.8 || die
 }
